Change a few floats to CGFloats. Make sure selection is visible when dragging. Attempt at making sure selection animates during tracking.

+++ ULIPaintView.h
@@ -46,8 +46,8 @@
 	NSTrackingRectTag		tempTrackTag;
 	NSRect					selectionFrame;
 	NSBezierPath*			selectionPath;
-	float					selectionAlpha;
-	float					selectionPhase;
+	CGFloat					selectionAlpha;
+	CGFloat					selectionPhase;
 	NSTimer*				selectionTimer;
 	float					selectionTimerDelta;
 	BOOL					drawSelectionHighlight;
